--- doonut-app-external-api/src/main/kotlin/com/doonutmate/example/controller/dto/TokenRequest.kt
@@ -1,8 +1,0 @@
-package com.doonutmate.example.controller.dto
-
-import jakarta.validation.constraints.NotEmpty
-
-data class TokenRequest(
-    @NotEmpty
-    val accessToken: String
-)
