+++ doonut-app-external-api/src/main/kotlin/com/doonutmate/oauth/configuration/KakaoOauthConfig.kt
@@ -1,0 +1,18 @@
+package com.doonutmate.oauth.configuration
+
+import org.springdoc.core.models.GroupedOpenApi
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class KakaoOauthConfig {
+    @Bean
+    fun getKakaoOauthApi(): GroupedOpenApi {
+        return GroupedOpenApi.builder()
+            .group("kakaoOauth")
+            .pathsToMatch("/oauth/**")
+            .pathsToExclude("")
+            .packagesToScan("com.doonutmate.oauth.controller")
+            .build()
+    }
+}
